Tidied R code, especially for plots
Made a dedicated subfolder for plots made externally to models. Made the plotting code run without exploratory code which looked to split results into the distributional parameters. This should be looked at later on for a proper summary of the data though. Moved testing/incomplete code to a new folder within R/