updated all plots, added error type plots
Updated all plots of the summary data and posterior summaries to use a consistent style. Also added a breakdown of how errors occur when they happen and breakdowns of the data by beta effects and zoi/coi effects. Renamed folders for plots and posterior summaries of results to have more obvious names.