Added all plots with and without titles
Changed code for all plot producing files so that plots are saved both with and without titles and added to the relevant subfolders. This required a change to ensure that in-plot annotations only applies to one facet in the data summary code for the ZOI/COI plots rather than forcing an annotation in cowplot. Thus, cowplot is no longer a dependency in this code.